+++ autodistill/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.1.15"
+__version__ = "0.1.16"
